refactor(studio): isolate clip drag lifecycle
